+++ app/Http/Controllers/ClientesController.php
@@ -20,6 +20,10 @@ class ClientesController extends Controller
 
     public function create()
     {
+        if (!verificarPermiso('Agregar_Clientes')) {
+            session()->flash('accesoDenegado');
+            return redirect()->back();
+        }
         return view('dashboard.clientes.create');
     }
 
@@ -66,12 +70,20 @@ class ClientesController extends Controller
 
     public function show(string $id)
     {
-
+        if (!verificarPermiso('Ver_Clientes')) {
+            session()->flash('accesoDenegado');
+            return redirect()->back();
+        }
     }
 
 
     public function edit($id)
     {
+        if (!verificarPermiso('Editar_Clientes')) {
+            session()->flash('accesoDenegado');
+            return redirect()->back();
+        }
+
         $url = env('URL_SERVER_API', 'http://127.0.0.1:8000');
         $response = Http::get($url.'/clientes/'.$id);
 
@@ -121,8 +133,12 @@ class ClientesController extends Controller
 
     public function destroy(string $id)
     {
-        $url = env('URL_SERVER_API', 'http://127.0.0.1:8000');
+        if (!verificarPermiso('Eliminar_Clientes')) {
+            session()->flash('accesoDenegado');
+            return redirect()->back();
+        }
 
+        $url = env('URL_SERVER_API', 'http://127.0.0.1:8000');
         $response = Http::delete($url.'/clientes/'.$id);
         $result = $response->json();
