+++ app/Http/Controllers/MarcaController.php
@@ -18,6 +18,11 @@ class MarcaController extends Controller
 
     public function create()
     {
+        if (!verificarPermiso('Agregar_Marcas')) {
+            session()->flash('accesoDenegado');
+            return redirect()->back();
+        }
+
         return view('dashboard.marcas.create');
     }
 
@@ -58,6 +63,11 @@ class MarcaController extends Controller
 
     public function edit(string $id)
     {
+        if (!verificarPermiso('Editar_Marcas')) {
+            session()->flash('accesoDenegado');
+            return redirect()->back();
+        }
+
         $url = env('URL_SERVER_API', 'http://127.0.0.1:8000');
         $response = Http::get($url . '/marcas/' . $id);
         $marca = $response->json();
@@ -95,6 +105,11 @@ class MarcaController extends Controller
 
     public function destroy(string $id)
     {
+        if (!verificarPermiso('Eliminar_Marcas')) {
+            session()->flash('accesoDenegado');
+            return redirect()->back();
+        }
+
         $url = env('URL_SERVER_API', 'http://127.0.0.1:8000');
         $response = Http::delete($url . '/marcas/' . $id);
         $result = $response->json();
