+++ app/Http/Controllers/RolesController.php
@@ -18,6 +18,11 @@ class RolesController extends Controller
 
     public function create()
     {
+        if (!verificarPermiso('Agregar_Roles')) {
+            session()->flash('accesoDenegado');
+            return redirect()->back();
+        }
+
         return view('dashboard.roles.create');
     }
 
@@ -54,12 +59,20 @@ class RolesController extends Controller
 
     public function show(string $id)
     {
-        //
+        if (!verificarPermiso('Ver_Roles')) {
+            session()->flash('accesoDenegado');
+            return redirect()->back();
+        }
     }
 
 
     public function edit(string $id)
     {
+        if (!verificarPermiso('Editar_Roles')) {
+            session()->flash('accesoDenegado');
+            return redirect()->back();
+        }
+
         $url = env('URL_SERVER_API', 'http://127.0.0.1:8000');
         $response = Http::get($url . '/roles/' . $id);
         $rol = $response->json();
@@ -98,6 +111,11 @@ class RolesController extends Controller
 
     public function destroy(string $id)
     {
+        if (!verificarPermiso('Eliminar_Roles')) {
+            session()->flash('accesoDenegado');
+            return redirect()->back();
+        }
+
         $url = env('URL_SERVER_API', 'http://127.0.0.1:8000');
         $response = Http::delete($url . '/roles/' . $id);
         $result = $response->json();
