+++ app/Http/Controllers/VehiculoController.php
@@ -18,6 +18,11 @@ class VehiculoController extends Controller
 
     public function create()
     {
+        if (!verificarPermiso('Agregar_Vehiculos')) {
+            session()->flash('accesoDenegado');
+            return redirect()->back();
+        }
+
         $url = env('URL_SERVER_API', 'http://127.0.0.1:8000');
         $responseMarcas = Http::get($url . '/marcas');
         $responseModelos = Http::get($url . '/modelos');
@@ -76,12 +81,20 @@ class VehiculoController extends Controller
 
     public function show(string $id)
     {
-        //
+        if (!verificarPermiso('Ver_Vehiculos')) {
+            session()->flash('accesoDenegado');
+            return redirect()->back();
+        }
     }
 
 
     public function edit(string $id)
     {
+        if (!verificarPermiso('Editar_Vehiculos')) {
+            session()->flash('accesoDenegado');
+            return redirect()->back();
+        }
+
         $url = env('URL_SERVER_API', 'http://127.0.0.1:8000');
         $responseVehiculo = Http::get($url . '/vehiculos/' . $id);
         $responseMarcas = Http::get($url . '/marcas');
@@ -142,6 +155,11 @@ class VehiculoController extends Controller
 
     public function destroy($id)
     {
+        if (!verificarPermiso('Eliminar_Vehiculos')) {
+            session()->flash('accesoDenegado');
+            return redirect()->back();
+        }
+
         $url = env('URL_SERVER_API', 'http://127.0.0.1:8000');
         $response = Http::delete($url . '/vehiculos/' . $id);
         $result = $response->json();
