+++ app/Http/Controllers/proveedorController.php
@@ -17,6 +17,11 @@ class proveedorController extends Controller
 
     public function create()
     {
+        if (!verificarPermiso('Agregar_Proveedores')) {
+            session()->flash('accesoDenegado');
+            return redirect()->back();
+        }
+
         return view('dashboard.proveedor.create');
     }
 
@@ -54,6 +59,11 @@ class proveedorController extends Controller
 
     public function edit($id)
     {
+        if (!verificarPermiso('Editar_Proveedores')) {
+            session()->flash('accesoDenegado');
+            return redirect()->back();
+        }
+
         $url = env('URL_SERVER_API', 'http://127.0.0.1:8000');
         $response = Http::get($url . '/proveedores/' . $id);
         $proveedor = $response->json();
@@ -93,6 +103,11 @@ class proveedorController extends Controller
 
     public function destroy($id)
     {
+        if (!verificarPermiso('Eliminar_Proveedores')) {
+            session()->flash('accesoDenegado');
+            return redirect()->back();
+        }
+
         $url = env('URL_SERVER_API', 'http://127.0.0.1:8000');
         $response = Http::delete($url . '/proveedores/' . $id);
         $result = $response->json();
